Changed how registers work.
I had 80+ errors, WHY GOD, WHY??? Sorry if I'm offending people of other
religions, please don't hate me :)

+++ ChronosVM 2/Program.cs
@@ -12,7 +12,7 @@ namespace ChronosVM_2
 {
     class Program
     {
-        public static int instructionSize = 10;
+        public static ushort instructionSize = 10;
 
         [DllImport("kernel32.dll")]
         static extern bool AttachConsole(int dwProcessId);
